backend/pyproject.toml
[project]
name = "sentiment-backend"
version = "0.1.0"
description = "Advisor Sentiment — FastAPI + SQLAlchemy async backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9",
    "sqlalchemy[asyncio]>=2.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "pydantic[email]>=2.0",
    "structlog>=24.0",
    "httpx>=0.27",
    "faster-whisper>=1.0",
    "aioboto3>=13.0",
    "opensearch-py[async]>=2.0",
    "apscheduler>=3.10",
    "redis>=5.0",
    "prometheus-client>=0.22",
    # OpenTelemetry
    "opentelemetry-sdk>=1.25",
    "opentelemetry-exporter-otlp-proto-http>=1.25",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "opentelemetry-instrumentation-httpx>=0.46b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.46b0",
]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "ruff>=0.15",
    "mypy>=1.0",
]

# ---------------------------------------------------------------------------
# Build system — minimal (app, not a distributable library)
# ---------------------------------------------------------------------------
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"

# ---------------------------------------------------------------------------
# ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = [
    "B008",  # FastAPI Depends() calls in function defaults
]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
warn_unused_ignores = false
